+++ src/Concerto/PanelBundle/Resources/starter_content/src/template/page/page.js
@@ -1,0 +1,3 @@
+testRunner.controllerProvider.register("page", function($scope) {
+  $scope.buttonLabel = testRunner.R.buttonLabel ? testRunner.R.buttonLabel : "Next";
+});